backend/requirements.txt
django==5.0.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
celery==5.4.0
redis==5.0.8
pylint==3.2.7
pylint-django==2.5.5
watchdog==5.0.2
colorlog==6.8.2
pdfplumber
# langchain
# langchain-google-genai
# langchain_community
python-dotenv
google-generativeai
azure-cognitiveservices-speech
moviepy
pillow
aiohttp
djangorestframework-simplejwt
django-allauth
dj-rest-auth
textract
youtube-transcript-api
typing
